AI Capabilities/LLM/requirements.txt
# llm_chat_service/requirements.txt

fastapi
uvicorn
pydantic
python-dotenv
pdfplumber
transformers
torch
langchain
langchain-core
langchain-community
langchain-mistralai
pinecone-client
langchain-pinecone